+++ packages/contracts/contracts/marketprotocol/MarketContractRegistry.sol
@@ -16,8 +16,8 @@
 
 pragma solidity 0.5.2;
 
-import 'openzeppelin-solidity/contracts/ownership/Ownable.sol';
-import './MarketContractRegistryInterface.sol';
+import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
+import "./MarketContractRegistryInterface.sol";
 
 
 /// @title MarketContractRegistry
@@ -54,14 +54,14 @@ contract MarketContractRegistry is Ownable, MarketContractRegistryInterface {
     /// a decentralized smart contract of community members to vote
     /// @param contractAddress contract to removed from white list
     /// @param whiteListIndex of the contractAddress in the addressWhiteList to be removed.
-    function removeContractFromWhiteList(address contractAddress, uint whiteListIndex)
+    function removeContractFromWhiteList(address contractAddress, uint256 whiteListIndex)
         external
         onlyOwner
     {
-        require(isWhiteListed[contractAddress], 'can only remove whitelisted addresses');
+        require(isWhiteListed[contractAddress], "can only remove whitelisted addresses");
         require(
             addressWhiteList[whiteListIndex] == contractAddress,
-            'index does not match address'
+            "index does not match address"
         );
         isWhiteListed[contractAddress] = false;
 
@@ -77,9 +77,9 @@ contract MarketContractRegistry is Ownable, MarketContractRegistryInterface {
     function addAddressToWhiteList(address contractAddress) external {
         require(
             isOwner() || factoryAddressWhiteList[msg.sender],
-            'Can only be added by factory or owner'
+            "Can only be added by factory or owner"
         );
-        require(!isWhiteListed[contractAddress], 'Address must not be whitelisted');
+        require(!isWhiteListed[contractAddress], "Address must not be whitelisted");
         isWhiteListed[contractAddress] = true;
         addressWhiteList.push(contractAddress);
         emit AddressAddedToWhitelist(contractAddress);
@@ -88,7 +88,7 @@ contract MarketContractRegistry is Ownable, MarketContractRegistryInterface {
     /// @dev allows for the owner to add a new address of a factory responsible for creating new market contracts
     /// @param factoryAddress address of factory to be allowed to add contracts to whitelist
     function addFactoryAddress(address factoryAddress) external onlyOwner {
-        require(!factoryAddressWhiteList[factoryAddress], 'address already added');
+        require(!factoryAddressWhiteList[factoryAddress], "address already added");
         factoryAddressWhiteList[factoryAddress] = true;
         emit FactoryAddressAdded(factoryAddress);
     }
@@ -98,7 +98,7 @@ contract MarketContractRegistry is Ownable, MarketContractRegistryInterface {
     function removeFactoryAddress(address factoryAddress) external onlyOwner {
         require(
             factoryAddressWhiteList[factoryAddress],
-            'factory address is not in the white list'
+            "factory address is not in the white list"
         );
         factoryAddressWhiteList[factoryAddress] = false;
         emit FactoryAddressRemoved(factoryAddress);
